moves elements into seperate components
